test: mv to to_string_test.go
